Extract ListSource and StackSource common functionalities to CollectionSource
